requirements2.txt
absl-py==1.4.0
alembic==1.14.0
altgraph==0.17.2
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.8.2
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.3.0
cajarename==19.7.15
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
colcon-argcomplete==0.3.3
colcon-bash==0.4.2
colcon-cd==0.1.1
colcon-cmake==0.2.27
colcon-common-extensions==0.3.0
colcon-core==0.10.0
colcon-defaults==0.2.6
colcon-devtools==0.2.3
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.14
colcon-output==0.2.12
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.2.4
colcon-pkg-config==0.1.0
colcon-powershell==0.3.7
colcon-python-setup-py==0.2.7
colcon-recursive-crawl==0.2.1
colcon-ros==0.3.23
colcon-test-result==0.3.8
colcon-zsh==0.4.0
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==4.5.2
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
deja-dup-caja==0.0.6
Deprecated==1.2.7
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
filelock==3.11.0
flake8==3.7.9
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.3.3
folder-color-caja==0.0.86
folder-color-common==0.0.86
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg===1.13.1-unknown
greenlet==3.1.1
grpcio @ https://files.pythonhosted.org/packages/06/1f/9f9cb7b5f1402122ad2a249dddbe7077507759a380808345841642a25c80/grpcio-1.54.0-cp38-cp38-manylinux_2_17_aarch64.whl#sha256=b7655f809e3420f80ce3bf89737169a9dce73238af594049754a1128132c0da4
h5py==3.8.0
html5lib==1.0.1
httplib2==0.14.0
humanfriendly==10.0
hyperlink==19.0.0
idna==3.4
ifcfg==0.18
importlib-metadata==6.4.1
importlib_resources==6.4.5
incremental==16.10.1
iso8601==1.1.0
itsdangerous==2.2.0
jax==0.4.8
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.12.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lark-parser==0.8.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
Magnus==1.0.3
Mako==1.1.0
Markdown==3.4.3
MarkupSafe==2.1.2
mate-hud==19.10.0
mate-menu==20.4.1
mate-tweak==20.4.0
matplotlib==3.1.2
mccabe==0.6.1
ml-dtypes==0.1.0
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
nose2==0.9.1
notify2==0.3
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
onnxruntime==1.14.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.1
PAM==0.4.2
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==9.5.0
pluggy==0.13.0
protobuf==4.22.3
psutil==5.5.1
pulsemixer==1.5.0
py==1.8.1
py-ubjson==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyassimp==4.1.4
PyBluez==0.23
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydocstyle==2.1.1
pydot==1.4.1
pyflakes==2.1.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyJWT==1.7.1
pymacaroons==0.13.0
pymycobot==3.1.3
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyreadline3==3.4.1
pyrsistent==0.15.5
pyserial==3.5
PySide6 @ file:///home/er/PySide6/PySide6-6.5.1.1-cp37-abi3-manylinux_2_31_aarch64.whl#sha256=b98a61508d2f5db68600ab0323af810e6cfe3848065fc62a5759ebb73ecb6159
PySide6-Addons @ file:///home/er/PySide6/PySide6_Addons-6.5.1.1-cp37-abi3-manylinux_2_31_aarch64.whl#sha256=5cf2f9c69cc3b2dce0f6e6162f89897db5ec22620acffbc91f8f6138363454d1
PySide6-Essentials @ file:///home/er/PySide6/PySide6_Essentials-6.5.1.1-cp37-abi3-manylinux_2_31_aarch64.whl#sha256=6549c6f9cab38e051ec3284536f0975fb8222f1be918adef63ac2c66bbe1a921
pytest==4.6.9
pytest-cov==2.8.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-snappy==0.5.3
python-xapp==1.8.1
python-xlib==0.23
PyTrie==0.2
pyxattr==0.6.1
pyxdg==0.26
PyYAML==6.0
reportlab==3.5.34
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosinstall==0.7.8
rosinstall-generator==0.1.22
rospkg==1.4.0
rospkg-modules==1.4.0
RPi.GPIO==0.7.1
rsa==4.9
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
shiboken6 @ file:///home/er/PySide6/shiboken6-6.5.1.1-cp37-abi3-manylinux_2_31_aarch64.whl#sha256=62a7c41fa6d5e609782e421018335bd010ecab4674c7700814849c357413eb8f
shiboken6-generator @ file:///home/er/PySide6/shiboken6_generator-6.5.1.1-cp37-abi3-manylinux_2_31_aarch64.whl#sha256=ee23cb43bf90bfba7b16a1b7f33e13202c8bf304df5063262cf4486a6def5ffa
simplejson==3.16.0
sip==4.19.21
six==1.16.0
snowballstemmer==2.0.0
soupsieve==1.9.5
SQLAlchemy==2.0.36
ssh-import-id==5.10
sympy==1.11.1
systemd-python==234
termcolor==2.2.0
tinyaes==1.0.3
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.12.2
u-msgpack-python==2.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.14.1
wsaccel==0.6.2
wstool==0.1.18
WTForms==3.1.2
xkit==0.0.0
xlwt==1.3.0
youtube-dl==2020.3.24
zipp==3.15.0
zope.interface==4.7.1
